requirements.txt
Django==2.0.13
django-cors-middleware==1.4.0
django-filter==2.2.0
django-rest-framework-roles==0.6
djangorestframework==3.10.2
Markdown==3.1.1
mysqlclient==1.4.4
pytz==2019.2
